requirements.txt
Flask
python-dotenv
Authlib
redis
requests
pandas
numpy
ta
psycopg2-binary
websockets
protobuf
joblib
lightgbm
scikit-learn
SQLAlchemy
pytz
python-dateutil
upstox-python-sdk
